refactor: 예약 상태 파일 이동
