+++ app/src/main/java/com/example/team25/ui/reservation/ReservationInfoViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import com.example.team25.domain.Gender
 import com.example.team25.domain.model.Patient
 import com.example.team25.domain.model.ReservationInfo
-import com.example.team25.domain.model.ReservationStatus
+import com.example.team25.domain.ReservationStatus
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
